Be lenient with army stats reported as empty array
LUA's JSON serializer has no way of knowing whether an empty table
should be serialized as an empty table or an empty object. As it chooses
empty array for 'blueprints' which should in fact be an object (so it
can be deserialized as Map), we need to detect this and convert the
empty array into an empty Map.