Fix 'Sleep is for the weak' achievement tests
TimeZone.getDefault() gets the host's timezone, however the parsed
Instant's were in UTC (that's what the Z at the end means). So these
tests would pass/fail depending on whether or not the host's timezone
is between UTC+[0, 3]. The TooEarly test was passing because it didn't
report any army results and never got to the time check anyways.